Refactor monkey tests for better reproducibility
The OutGameMonkeyTest and InGameMonkeyTest in the RoguelikeExample runtime integration tests have been reorganized for better reproducibility during testing.

The tests now utilize a pseudo-random number generator, and the seed value is logged for future reproduction of the tests. For scenes with random elements, this approach proves more convenient, as it only requires a single seed value. The test for the dungeon scene has been modified to allow the dungeonManager's seed value to be set, which simplifies the reproduction process. Removed any possible redundancy regarding the loading of the 'Title' and 'Dungeon' scenes.